requirements.txt
tensorflow>=2.16.0
numpy>=1.21.0
Pillow>=8.0.0
